Classify tests in folders
