+++ tests/LiveChartsCore.UnitTesting/CoreObjectsTests/ActionThrottler.cs
@@ -26,7 +26,7 @@ using System.Threading.Tasks;
 using LiveChartsCore.Kernel;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace LiveChartsCore.UnitTesting;
+namespace LiveChartsCore.UnitTesting.CoreObjectsTests;
 
 [TestClass]
 public class ActionThrottlerTester
