+++ tests/LiveChartsCore.UnitTesting/CoreObjectsTests/ColorTesting.cs
@@ -23,7 +23,7 @@
 using LiveChartsCore.Drawing;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace LiveChartsCore.UnitTesting;
+namespace LiveChartsCore.UnitTesting.CoreObjectsTests;
 
 [TestClass]
 public class ColorTesting
