+++ tests/LiveChartsCore.UnitTesting/SeriesTests/PieSeriesTest.cs
@@ -26,7 +26,7 @@ using LiveChartsCore.SkiaSharpView;
 using LiveChartsCore.SkiaSharpView.SKCharts;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace LiveChartsCore.UnitTesting;
+namespace LiveChartsCore.UnitTesting.Series;
 
 [TestClass]
 public class PieSeriesTest
