+++ tests/LiveChartsCore.UnitTesting/SeriesTests/ScatterSeriesTest.cs
@@ -22,21 +22,40 @@
 
 using System;
 using System.Linq;
+using LiveChartsCore.Defaults;
 using LiveChartsCore.SkiaSharpView;
 using LiveChartsCore.SkiaSharpView.SKCharts;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace LiveChartsCore.UnitTesting;
+namespace LiveChartsCore.UnitTesting.Series;
 
 [TestClass]
-public class RowSeriesTest
+public class ScatterSeriesTest
 {
     [TestMethod]
     public void ShouldScaleProperly()
     {
-        var sutSeries = new RowSeries<double>
+        var values = new WeightedPoint[]
         {
-            Values = new double[] { 1, 2, 4, 8, 16, 32, 64, 128, 256, 512 }
+            new(0, 1, 0),
+            new(1, 2, 1),
+            new(2, 4, 2),
+            new(3, 8, 3),
+            new(4, 16, 4),
+            new(5, 32, 5),
+            new(6, 64, 6),
+            new(7, 128, 7),
+            new(8, 256, 8),
+            new(9, 512, 9)
+        };
+
+        var maxW = values.Max(x => x.Weight)!.Value;
+
+        var sutSeries = new ScatterSeries<WeightedPoint>
+        {
+            Values = values,
+            MinGeometrySize = 10,
+            GeometrySize = 100
         };
 
         var chart = new SKCartesianChart
@@ -44,8 +63,8 @@ public class RowSeriesTest
             Width = 1000,
             Height = 1000,
             Series = new[] { sutSeries },
-            XAxes = new[] { new Axis { MinLimit = 0, MaxLimit = 512 } },
-            YAxes = new[] { new Axis { MinLimit = -1, MaxLimit = 10 } }
+            XAxes = new[] { new Axis { MinLimit = -1, MaxLimit = 10 } },
+            YAxes = new[] { new Axis { MinLimit = 0, MaxLimit = 512 } }
         };
 
         _ = chart.GetImage();
@@ -61,35 +80,37 @@ public class RowSeriesTest
 
         // ensure the unit has valid dimensions
         Assert.IsTrue(typedUnit.Visual.Width > 1 && typedUnit.Visual.Height > 1);
+        Assert.IsTrue(typedUnit.Visual.Width == sutSeries.MinGeometrySize && typedUnit.Visual.Height == sutSeries.MinGeometrySize);
 
         var previous = typedUnit;
-        float? previousY = null;
+        float? previousX = null;
 
         foreach (var sutPoint in toCompareGuys)
         {
-            // test width
-            Assert.IsTrue(
-                // the idea is, the second bar should be 2 times bigger than the first one
-                // and the third bar should be 4 times bigger than the first one and so on
-                Math.Abs(typedUnit.Visual.Width - sutPoint.Visual.Width / (float)sutPoint.Model) < 0.001);
+            var w = sutPoint.Model.Weight!.Value / maxW;
+            var targetSize = sutSeries.MinGeometrySize + (sutSeries.GeometrySize - sutSeries.MinGeometrySize) * w;
 
             // test height
-            Assert.IsTrue(
-                // and also the width should be the same.
-                Math.Abs(typedUnit.Visual.Height - sutPoint.Visual.Height) < 0.001);
+            Assert.IsTrue(Math.Abs(sutPoint.Visual.Height - targetSize) < 0.001);
 
-            // test y
-            var currentDeltaY = previous.Visual.Y - sutPoint.Visual.Y;
-            Assert.IsTrue(
-                previousY is null
-                ||
-                Math.Abs(previousY.Value - currentDeltaY) < 0.001);
+            // test width
+            Assert.IsTrue(Math.Abs(sutPoint.Visual.Width - targetSize) < 0.001);
 
             // test x
+            var currentDeltaX = previous.Visual.X - sutPoint.Visual.X;
             Assert.IsTrue(
-                Math.Abs(sutPoint.Visual.X - chart.Core.DrawMarginLocation.X) < 0.001);
+                previousX is null
+                ||
+                Math.Abs(previousX.Value - currentDeltaX) < 0.001);
 
-            previousY = previous.Visual.Y - sutPoint.Visual.Y;
+            // test y
+            var p = 1f - sutPoint.PrimaryValue / 512f;
+            Assert.IsTrue(
+                Math.Abs(
+                    p * chart.Core.DrawMarginSize.Height - sutPoint.Visual.Y -
+                    sutPoint.Visual.Height * 0.5f + chart.Core.DrawMarginLocation.Y) < 0.001);
+
+            previousX = previous.Visual.X - sutPoint.Visual.X;
             previous = sutPoint;
         }
     }
