+++ tests/LiveChartsCore.UnitTesting/SeriesTests/StackedAreaSeriesTest.cs
@@ -26,17 +26,23 @@ using LiveChartsCore.SkiaSharpView;
 using LiveChartsCore.SkiaSharpView.SKCharts;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace LiveChartsCore.UnitTesting;
+namespace LiveChartsCore.UnitTesting.Series;
 
 [TestClass]
-public class StepLineSeriesTest
+public class StackedAreaSeriesTest
 {
     [TestMethod]
     public void ShouldScaleProperly()
     {
-        var sutSeries = new StepLineSeries<double>
+        var sutSeries = new StackedAreaSeries<double>
         {
-            Values = new double[] { 1, 2, 4, 8, 16, 32, 64, 128, 256, 512 },
+            Values = new double[] { 1, 2, 4, 8, 16, 32, 64, 128, 256 },
+            GeometrySize = 10
+        };
+
+        var sutSeries2 = new StackedAreaSeries<double>
+        {
+            Values = new double[] { 1, 2, 4, 8, 16, 32, 64, 128, 256 },
             GeometrySize = 10
         };
 
@@ -44,13 +50,13 @@ public class StepLineSeriesTest
         {
             Width = 1000,
             Height = 1000,
-            Series = new[] { sutSeries },
+            Series = new[] { sutSeries, sutSeries2 },
             XAxes = new[] { new Axis { MinLimit = -1, MaxLimit = 10 } },
             YAxes = new[] { new Axis { MinLimit = 0, MaxLimit = 512 } }
         };
 
         _ = chart.GetImage();
-        // chart.SaveImage("test.png"); // use this method to see the actual tested image
+        //chart.SaveImage("test.png"); // use this method to see the actual tested image
 
         var datafactory = sutSeries.DataFactory;
         var points = datafactory.Fetch(sutSeries, chart.Core).ToArray();
@@ -59,6 +65,12 @@ public class StepLineSeriesTest
         var typedUnit = sutSeries.ConvertToTypedChartPoint(unit);
 
         var toCompareGuys = points.Where(x => x != unit).Select(sutSeries.ConvertToTypedChartPoint);
+
+        var datafactory2 = sutSeries2.DataFactory;
+        var points2 = datafactory2.Fetch(sutSeries2, chart.Core).ToArray();
+        var unit2 = points2.First(x => x.PrimaryValue == 1);
+        var typedUnit2 = sutSeries.ConvertToTypedChartPoint(unit2);
+        var toCompareGuys2 = points2.Where(x => x != unit2).Select(sutSeries2.ConvertToTypedChartPoint);
 
         // ensure the unit has valid dimensions
         Assert.IsTrue(typedUnit.Visual.Geometry.Width == 10 && typedUnit.Visual.Geometry.Height == 10);
@@ -71,7 +83,7 @@ public class StepLineSeriesTest
         {
             // test x
             var currentDeltaX = previous.Visual.Geometry.X - sutPoint.Visual.Geometry.X;
-            var currentDeltaAreaX = previous.Visual.StepSegment.Xj - sutPoint.Visual.StepSegment.Xj;
+            var currentDeltaAreaX = previous.Visual.Bezier.Xj - sutPoint.Visual.Bezier.Xj;
             Assert.IsTrue(
                 previousX is null
                 ||
@@ -82,14 +94,43 @@ public class StepLineSeriesTest
                 Math.Abs(previousXArea.Value - currentDeltaX) < 0.001);
 
             // test y
-            var p = 1f - sutPoint.PrimaryValue / 512f;
+            var p = 1f - (sutPoint.PrimaryValue + sutPoint.StackedValue.Start) / 512f;
             Assert.IsTrue(
                 Math.Abs(p * chart.Core.DrawMarginSize.Height - sutPoint.Visual.Geometry.Y + chart.Core.DrawMarginLocation.Y) < 0.001);
             Assert.IsTrue(
-                Math.Abs(p * chart.Core.DrawMarginSize.Height - sutPoint.Visual.StepSegment.Yj + chart.Core.DrawMarginLocation.Y) < 0.001);
+                Math.Abs(p * chart.Core.DrawMarginSize.Height - sutPoint.Visual.Bezier.Yj + chart.Core.DrawMarginLocation.Y) < 0.001);
 
             previousX = previous.Visual.Geometry.X - sutPoint.Visual.Geometry.X;
-            previousXArea = previous.Visual.StepSegment.Xj - sutPoint.Visual.StepSegment.Xj;
+            previousXArea = previous.Visual.Bezier.Xj - sutPoint.Visual.Bezier.Xj;
+            previous = sutPoint;
+        }
+
+        previous = typedUnit2;
+        previousX = null;
+        previousXArea = null;
+        foreach (var sutPoint in toCompareGuys2)
+        {
+            // test x
+            var currentDeltaX = previous.Visual.Geometry.X - sutPoint.Visual.Geometry.X;
+            var currentDeltaAreaX = previous.Visual.Bezier.Xj - sutPoint.Visual.Bezier.Xj;
+            Assert.IsTrue(
+                previousX is null
+                ||
+                Math.Abs(previousX.Value - currentDeltaX) < 0.001);
+            Assert.IsTrue(
+                previousXArea is null
+                ||
+                Math.Abs(previousXArea.Value - currentDeltaX) < 0.001);
+
+            // test y
+            var p = 1f - (sutPoint.PrimaryValue + sutPoint.StackedValue.Start) / 512f;
+            Assert.IsTrue(
+                Math.Abs(p * chart.Core.DrawMarginSize.Height - sutPoint.Visual.Geometry.Y + chart.Core.DrawMarginLocation.Y) < 0.001);
+            Assert.IsTrue(
+                Math.Abs(p * chart.Core.DrawMarginSize.Height - sutPoint.Visual.Bezier.Yj + chart.Core.DrawMarginLocation.Y) < 0.001);
+
+            previousX = previous.Visual.Geometry.X - sutPoint.Visual.Geometry.X;
+            previousXArea = previous.Visual.Bezier.Xj - sutPoint.Visual.Bezier.Xj;
             previous = sutPoint;
         }
     }
